+++ packages/cli/src/cli/folder.ts
@@ -5,7 +5,7 @@ import { CogJob } from '../cog/types';
 
 /** Make a temp folder inside TEMP_FOLDER's path */
 export async function makeTempFolder(folder: string): Promise<string> {
-    const tempPath = Env.get(Env.TempFolder, '/tmp');
+    const tempPath = Env.get(Env.TempFolder) ?? '/tmp';
     const folderPath = path.join(tempPath, folder);
 
     await fs.mkdir(folderPath, { recursive: true });