+++ packages/lambda-api-tracker/src/index.ts
@@ -11,10 +11,23 @@ function setTileInfo(ctx: LambdaContext): void {
         ctx.set('xyz', { x, y, z });
         ctx.set('projection', xyzData.projection.code);
         ctx.set('extension', ext);
+        ctx.set('tileSet', xyzData.name);
 
         const latLon = ProjectionTileMatrixSet.get(xyzData.projection.code).tileCenterToLatLon(xyzData);
         ctx.set('location', latLon);
     }
+}
+
+/** Extract the hostname from a url */
+export function getUrlHost(ref: string | undefined): string | undefined {
+    if (ref == null) return ref;
+    try {
+        const url = new URL(ref);
+        if (url.hostname) return url.hostname;
+    } catch (e) {
+        // Ignore
+    }
+    return ref;
 }
 
 /**
@@ -27,7 +40,11 @@ export async function handleRequest(req: LambdaContext): Promise<LambdaHttpRespo
         req.set('clientIp', req.evt.Records[0].cf.request.clientIp);
     }
 
-    req.set('referer', req.header('referer'));
+    /**
+     * We generally don't need the full referer to get
+     * a understanding of where the request came from
+     */
+    req.set('referer', getUrlHost(req.header('referer')));
     req.set('userAgent', req.header('user-agent'));
 
     if (req.action.name === 'tiles') setTileInfo(req);