+++ packages/lambda-xyz/src/png.ts
@@ -1,4 +1,0 @@
-export const EmptyPng = Buffer.from(
-    `iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAAC0lEQVR4nGNgAAIAAAUAAXpeqz8=`,
-    'base64',
-);