+++ packages/lambda-xyz/src/routes/api.ts
@@ -1,4 +1,3 @@
-import { Env } from '@basemaps/shared';
 import { LambdaHttpResponse } from '@basemaps/lambda';
 
 export async function Health(): Promise<LambdaHttpResponse> {
@@ -11,6 +10,6 @@ export async function Ping(): Promise<LambdaHttpResponse> {
 
 export async function Version(): Promise<LambdaHttpResponse> {
     const response = new LambdaHttpResponse(200, 'ok');
-    response.json({ version: process.env[Env.Version], hash: process.env[Env.Hash] });
+    response.json({ version: process.env.GIT_VERSION, hash: process.env.GIT_HASH });
     return response;
 }