+++ packages/linzjs-geojson/src/types.ts
@@ -1,0 +1,3 @@
+export { MultiPolygon, Polygon, Ring, Pair } from 'polygon-clipping';
+
+export type BBox = [number, number, number, number];