+++ packages/shared/src/proj/tile.set.name.ts
@@ -1,0 +1,7 @@
+export enum TileSetName {
+    aerial = 'aerial',
+}
+
+export function TileSetNameValues(): TileSetName[] {
+    return [TileSetName.aerial];
+}