+++ CHANGELOG.md
@@ -3,6 +3,43 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.5.0](https://github.com/linz/basemaps/compare/v4.4.0...v4.5.0) (2020-07-30)
+
+
+### Bug Fixes
+
+* **cli:** correctly detect if a tiff file list is passed in ([#993](https://github.com/linz/basemaps/issues/993)) ([9147c8e](https://github.com/linz/basemaps/commit/9147c8e8a6264dab285d6ab4a646f09d9a2d7718))
+* **cli:** folders must be mounted to allow docker to read the source files ([#995](https://github.com/linz/basemaps/issues/995)) ([8557afa](https://github.com/linz/basemaps/commit/8557afaf96c5153cd174d0af5475f1d2b3fe3f98))
+
+
+### Features
+
+* **cli:** support giving exact list of files to use ([#986](https://github.com/linz/basemaps/issues/986)) ([63b34ff](https://github.com/linz/basemaps/commit/63b34ff5414989b5d014b6d61f9be304ebd9e1e1))
+* **cli:** support guessing of CITM projection from wkt ([#994](https://github.com/linz/basemaps/issues/994)) ([61bb93b](https://github.com/linz/basemaps/commit/61bb93b23e8e02bf3cb7c1e983bbb64c6fcb3044))
+* **doc:** added examples used for snippets ([#912](https://github.com/linz/basemaps/issues/912)) ([3726f9e](https://github.com/linz/basemaps/commit/3726f9e3220f6426836e62ac5eeb3ced7e01b6d0))
+
+
+
+
+
+# [4.4.0](https://github.com/linz/basemaps/compare/v4.3.0...v4.4.0) (2020-07-28)
+
+
+### Bug Fixes
+
+* inject git version and hash into all code ([#966](https://github.com/linz/basemaps/issues/966)) ([8b8eaec](https://github.com/linz/basemaps/commit/8b8eaec373286c81b425d485274edd7c588aefea))
+* **lambda-api:** track api key usage ([#943](https://github.com/linz/basemaps/issues/943)) ([7c4689c](https://github.com/linz/basemaps/commit/7c4689cd0824ee678260ba5d84b25042aad72363))
+
+
+### Features
+
+* **lambda-api:** validate that the api key looks like a api key ([#954](https://github.com/linz/basemaps/issues/954)) ([badca1e](https://github.com/linz/basemaps/commit/badca1e904097a9275e6dbac576c010fccbfa345))
+* **lambda-xyz:** Serve WMTSCapabilities for all TileSets ([#953](https://github.com/linz/basemaps/issues/953)) ([49d0e88](https://github.com/linz/basemaps/commit/49d0e881b4726188ea937a9617c98bff5a78e44d))
+
+
+
+
+
 # [4.3.0](https://github.com/linz/basemaps/compare/v4.2.0...v4.3.0) (2020-07-19)