+++ packages/lambda-api-tracker/CHANGELOG.md
@@ -3,6 +3,31 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.5.0](https://github.com/linz/basemaps/compare/v4.4.0...v4.5.0) (2020-07-30)
+
+**Note:** Version bump only for package @basemaps/lambda-api-tracker
+
+
+
+
+
+# [4.4.0](https://github.com/linz/basemaps/compare/v4.3.0...v4.4.0) (2020-07-28)
+
+
+### Bug Fixes
+
+* inject git version and hash into all code ([#966](https://github.com/linz/basemaps/issues/966)) ([8b8eaec](https://github.com/linz/basemaps/commit/8b8eaec373286c81b425d485274edd7c588aefea))
+* **lambda-api:** track api key usage ([#943](https://github.com/linz/basemaps/issues/943)) ([7c4689c](https://github.com/linz/basemaps/commit/7c4689cd0824ee678260ba5d84b25042aad72363))
+
+
+### Features
+
+* **lambda-api:** validate that the api key looks like a api key ([#954](https://github.com/linz/basemaps/issues/954)) ([badca1e](https://github.com/linz/basemaps/commit/badca1e904097a9275e6dbac576c010fccbfa345))
+
+
+
+
+
 # [4.2.0](https://github.com/linz/basemaps/compare/v4.1.0...v4.2.0) (2020-07-16)
 
 **Note:** Version bump only for package @basemaps/lambda-api-tracker